Be/feat/160 (#173)
* refactor: MatchEvaluator refactoring

* feat: 경기 수정 기능 완료

* refactor: 경기 수정 리팩토링

* fix: patch games 유효성 검증 수정

* refactor: post api와 Request 통일

* refactor: post api와 Request 통일

* refactor: post api와 Request 통일

+++ be/xmash/src/main/java/com/tmp/xmash/web/GameController.java
@@ -56,7 +56,7 @@ public class GameController {
     @PatchMapping("/games/{matchId}")
     public ResponseEntity<Void> modifyGameResult(
             HttpSession session,
-            @Parameter(name = "matchId", description = "match_history_id") long matchId,
+            @Parameter(name = "matchId", description = "match_history_id") int matchId,
             @Parameter(name = "GameRequest", description = "게임 결과 등록 요청 데이터", required = true) @Valid @RequestBody GameResultRequest gameModifyRequest
     ) {
         String userId = (String) session.getAttribute("userId");
